+++ libminifi/src/FlowControlProtocol.cpp
@@ -193,7 +193,7 @@ int FlowControlProtocol::sendRegisterReq() {
   data = this->encode(data, this->_controller->getName());
 
   // send it
-  int status = sendData(buffer.data(), size);
+  int status = sendData(buffer.data(), gsl::narrow<int>(size));
   buffer.clear();
   if (status <= 0) {
     utils::file::FileUtils::close(_socket);
@@ -287,7 +287,7 @@ int FlowControlProtocol::sendReportReq() {
   data = this->encode(data, this->_controller->getName());
 
   // send it
-  int status = sendData(buffer.data(), size);
+  int status = sendData(buffer.data(), gsl::narrow<int>(size));
   buffer.clear();
   if (status <= 0) {
     utils::file::FileUtils::close(_socket);
